+++ tests/test_commandline.py
@@ -2,11 +2,12 @@ import subprocess
 import sys
 import os
 from io import StringIO, BytesIO
+from pathlib import Path
 
 import dnaio
 import pytest
 
-from cutadapt.__main__ import main
+from cutadapt.__main__ import main, CommandLineError
 from utils import assert_files_equal, datapath, cutpath
 
 # pytest.mark.timeout will not fail even if pytest-timeout is not installed
@@ -1044,6 +1045,23 @@ def test_rename_cannot_be_combined_with_other_renaming_options(opt):
                 datapath("empty.fastq"),
             ]
         )
+
+
+def test_duplicate_output_paths(tmp_path):
+    path = str(tmp_path / "discard.fastq")
+    with pytest.raises(SystemExit) as e:
+        main(
+            [
+                "--untrimmed-output",
+                path,
+                "--too-long-output",
+                path,
+                "-o",
+                os.devnull,
+                datapath("empty.fastq"),
+            ]
+        )
+    # "specified more than once as an output file"
 
 
 def test_rename(run):