+++ CHANGES.rst
@@ -6,6 +6,9 @@ Changelog
 development version
 -------------------
 
+* :issue:`607`: Print an error when an output file was specified
+  multiple times (for example, for ``--untrimmed-output`` and
+  ``--too-short-output``). This is not supported at the moment.
 * :issue:`603`: When ``-e`` was used with an absolute number of errors
   and there were ``N`` wildcards in the sequence, the actual number of
   allowed errors was too low.